# ADD - Plugin setting, EDIT - DBController, etc. 1. Plugin 구조에 맞추기 시작했습니다. 아직은 지속적으로 메시지가 들어오는것을 감시하는 등의 구현은 되어있지 않아서 단순히 존재하는 파일에서 읽어오는 것을 확인한 단계이지만, 알고리즘 로직의 구동 확인은 가능합니다.
2. DB 컨트롤러를 전반적으로 재설계했습니다. 기존의 db_lib에서 db 접근을
컨트롤했는데, 세션을 플러그인 쪽에서 관리해야하기 때문에 db_lib의 쿼리
실행과 같은 기능을 모두 db_controller로 통일했습니다.
커넥션은 pool 형태로 관리되며, config 파일에 풀의 크기 설정이 있습니다.

3. genesis-block 파일을 world-create로 변경하였습니다. 내부에 있는
정보들이 블록 자체에 대한 정보라기보다는 월드와 체인의 생성에 관한
내용이기 때문입니다.

4. first_block_test.json 파일을 추가하였습니다.
아직은 직접적으로 메시지 통신에 의한 데이터 처리를 테스트할 단계는
아니기에, 메시지의 형태를 가진 파일을 읽어오는 것으로 처리를 대체합니다.
트랜잭션의 txagg_cbor 관련 정보는 정확히 생성하지 않아서, 관련 정보를
처리하는 코드 또한 임시로 주석처리 되어있습니다. 다음 테스트때에는
트랜잭션에 대한 정보까지 정확하게 수정할 예정입니다.

5. block에서 user certification을 처리하는 부분을 수정하였습니다.
pem이 여러 개의 string 형태의 배열로 오기 때문에, 관련 처리를 해야
합니다.
pk 형식은 한 줄로 된 정보일거라고 하시기 때문에, 추후 pk/pem이 둘 다 올
수 있는 항목들에 대해서는 처리가 필요해 보입니다.

6. 로그가 기존의 easylog로 남아있는 것들이 많아서 public의 spdlog쪽의
함수로 변경했습니다. 단, 아직 완전한 정상 작동은 보장하지 않습니다.

7. levelDB를 추가하였습니다. 아직 Key-Value Store쪽은 구현되지
않았습니다.

+++ src/plugins/chain_plugin/include/unresolved_block_pool.hpp
@@ -6,6 +6,7 @@
 
 #include "../config/storage_type.hpp"
 #include "../../../../lib/gruut-utils/src/type_converter.hpp"
+#include "../../../../lib/log/include/log.hpp"
 
 #include "../structure/block.hpp"
 #include "block_validator.hpp"
@@ -16,8 +17,8 @@ namespace gruut {
 
 struct UnresolvedBlock {
   Block block;
-  int prev_vector_idx{-1};
-  size_t ssig_sum{0};
+  int32_t prev_vector_idx{-1};
+  int32_t ssig_sum{0};
   MemLedger m_mem_ledger;
 
   UnresolvedBlock() = default;
@@ -25,11 +26,11 @@ struct UnresolvedBlock {
 };
 
 struct BlockPosPool {
-  size_t height{0};
-  size_t vector_idx{0};
+  int32_t height{0};
+  int32_t vector_idx{0};
 
   BlockPosPool() = default;
-  BlockPosPool(size_t height_, size_t vector_idx_) : height(height_), vector_idx(vector_idx_) {}
+  BlockPosPool(int32_t height_, int32_t vector_idx_) : height(height_), vector_idx(vector_idx_) {}
 };
 
 class UnresolvedBlockPool {
@@ -54,7 +55,7 @@ private:
 public:
   UnresolvedBlockPool();
 
-  inline size_t size() {
+  inline int32_t size() {
     return m_block_pool.size();
   }
